Update InsertionSort.java
This is similar to a bubble sort, but technically sifts down values rather than bubbling them up.  Compared to a traditional bubble sort, it also iterates from the lowest index up but then sifts down as it reaches new values. While the overall iteration through the indices is pretty standard, the subsequent comparisons moving from the current index back down to 0 were not as clear since decrementing 'j' was done during the swap operation.  So my best recommendation to make this more clear, while maintaining the original logic, would be to move out the swap operation into its own method.  Then, decrement 'j' on its own line so that the downward movement of the comparison index is immediately apparent.